# api/templates/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MetaTrader 5 Flask REST API",
    "description": "Comprehensive REST API exposing 100% of the official Python MetaTrader 5 library methods in a headless Wine container.\n\n### Modules\n- **System**: MT5 lifecycle, connection, login, terminal and account status.\n- **Symbols & DOM**: Symbol listings, specifications, real-time ticks, and Depth of Market.\n- **Market Data**: Historical rates (OHLCV) and ticks with flexible filtering.\n- **Trading**: Market/pending order execution, positions management, margin & profit calculators.\n- **History**: Closed orders and deals history with date and ticket filters.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Local API Server"
    }
  ],
...
  "paths": {
...
    "/api/symbols_get": {
      "get": {
        "tags": [
          "Symbols & DOM"
        ],
        "summary": "Get Symbols List",
        "description": "Returns list of symbols with specifications. Can filter by group pattern (e.g. `*USD*`, `EUR*`, `*crypto*`).",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "description": "Group filter pattern",
            "required": false,
            "schema": {
              "type": "string",
              "default": "*"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of symbols",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/symbol_info/{symbol}": {
      "get": {
        "tags": [
          "Symbols & DOM"
        ],
        "summary": "Get Symbol Info",
        "description": "Returns complete financial instrument properties (point size, spread, digits, contract size, trade modes, margin rates).",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Symbol name (e.g., EURUSD, BTCUSD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Symbol specifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/symbol_info_tick/{symbol}": {
      "get": {
        "tags": [
          "Symbols & DOM"
        ],
        "summary": "Get Symbol Current Tick",
        "description": "Returns the most recent price tick (bid, ask, last, volume, time) for the specified symbol.",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Symbol name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest tick data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/symbol_select": {
      "post": {
        "tags": [
          "Symbols & DOM"
        ],
        "summary": "Select/Deselect Symbol in MarketWatch",
        "description": "Adds or removes a symbol from the Market Watch window.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymbolSelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Selection result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/market_book_add": {
      "post": {
        "tags": [
          "Symbols & DOM"
        ],
        "summary": "Subscribe to Depth of Market (DOM)",
        "description": "Subscribes the terminal to market book events for the symbol.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/market_book_get/{symbol}": {
      "get": {
        "tags": [
          "Symbols & DOM"
        ],
        "summary": "Get Depth of Market (DOM) Snapshots",
        "description": "Returns an array of DOM entries (buy/sell orders, prices, and volumes) for a subscribed symbol.",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Symbol name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DOM entries list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/market_book_release": {
      "post": {
        "tags": [
          "Symbols & DOM"
        ],
        "summary": "Unsubscribe from Depth of Market (DOM)",
        "description": "Cancels subscription to market book events for the symbol.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsubscription status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "SymbolSelectRequest": {
        "type": "object",
        "required": [
          "symbol"
        ],
        "properties": {
          "symbol": {
            "type": "string",
            "example": "EURUSD"
          },
          "enable": {
            "type": "boolean",
            "default": true,
            "description": "True to add, False to remove"
          }
        }
      },
      "MarketBookRequest": {
        "type": "object",
        "required": [
          "symbol"
        ],
        "properties": {
          "symbol": {
            "type": "string",
            "example": "EURUSD"
          }
        }
      },
...
    }
  }
}